[김민철] 진행 중 일 1. 사진 넣기

+++ refrigerator/src/main/java/moja/refrigerator/repository/recipe/RecipeSourceTypeRepository.java
@@ -3,5 +3,7 @@ package moja.refrigerator.repository.recipe;
 import moja.refrigerator.aggregate.recipe.RecipeSourceType;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface RecipeSourceTypeRepository extends JpaRepository<RecipeSourceType, Long> {
+import java.util.Optional;
+
+public interface RecipeSourceTypeRepository extends JpaRepository<RecipeSourceType, Integer> {
 }
